refactor(repair): move intensityHandler to separate file
